# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rifa API Docs",
        "description": "API para gerenciamento de rifas",
        "contact": {
            "name": "Felipe Silva",
            "email": "felipe.allware@gmail.com"
        },
        "license": {
            "name": "API Licence",
            "url": "https://github.com/felipesilva15/rifa-api/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/raffle": {
            "get": {
                "tags": [
                    "Raffle"
                ],
                "summary": "List all raffles",
                "operationId": "dcb7d24965c3e198389772664463eab2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Raffle ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Raffle name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "maximum_numbers",
                        "in": "query",
                        "description": "Raffle maximum numbers",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Raffle start date",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Raffle end date",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "ticket_value",
                        "in": "query",
                        "description": "Raffle ticket value",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Raffle list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Raffle"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Raffle"
                ],
                "summary": "Registers an raffle",
                "operationId": "5c0b16acd673b9206629c1757fe8abd5",
                "requestBody": {
                    "description": "Data for creating a new raffle",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Raffle"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered raffle data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Raffle"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/raffle/{id}": {
            "get": {
                "tags": [
                    "Raffle"
                ],
                "summary": "List an raffle by ID",
                "operationId": "8ba31a98e8774133f7b2bc8062827421",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Raffle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Raffle data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Raffle"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Raffle"
                ],
                "summary": "Update an raffle",
                "operationId": "3bbe011ce4ae881fcbf39ab208069e59",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Raffle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data for update raffle",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Raffle"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated raffle data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Raffle"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Raffle"
                ],
                "summary": "Deletes an raffle",
                "operationId": "bb457d470ff3ab5bfdcd8a9361ed9213",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Raffle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Response"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/raffle/{id}/tickets": {
            "get": {
                "tags": [
                    "Raffle"
                ],
                "summary": "List raffle tickets",
                "operationId": "6cb38e2c2df87ac255982a43f9ac35c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Raffle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/raffle/{id}/card": {
            "get": {
                "tags": [
                    "Raffle"
                ],
                "summary": "Get a raffle card",
                "operationId": "5978072b524793120522778ce3837d8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Raffle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Raffle card",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "number": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "avaiable": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error ocorried."
                    },
                    "route": {
                        "type": "string",
                        "example": "api/error"
                    }
                },
                "type": "object"
            },
...
            "Raffle": {
                "required": [
                    "name",
                    "maximum_numbers",
                    "start_date",
                    "ticket_value"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 3,
                        "example": "Xbox Series S"
                    },
                    "maximum_numbers": {
                        "type": "integer",
                        "example": 100
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-30"
                    },
                    "ticket_value": {
                        "type": "number",
                        "format": "double",
                        "maximum": 999.99,
                        "minimum": 0.01,
                        "example": 20
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30T03:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30T03:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "Ticket": {
                "required": [
                    "raffle_id",
                    "participant_id",
                    "number"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "raffle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "participant_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "number": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "payment_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-01"
                    },
                    "value": {
                        "type": "number",
                        "format": "double",
                        "maximum": 999.99,
                        "minimum": 0.01,
                        "example": 20
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30T03:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30T03:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}